Move TextEditorActivity to its own package
