+++ app/src/main/java/com/amaze/filemanager/asynchronous/asynctasks/SearchTextTask.java
@@ -26,7 +26,7 @@ import java.io.StringReader;
 import java.util.ArrayList;
 import java.util.Map;
 
-import com.amaze.filemanager.ui.activities.TextEditorActivity;
+import com.amaze.filemanager.ui.activities.texteditor.TextEditorActivity;
 import com.amaze.filemanager.ui.theme.AppTheme;
 import com.amaze.filemanager.utils.ImmutableEntry;
 import com.amaze.filemanager.utils.MapEntry;
